Split components and functions
