+++ client/app/hearings/components/modalForms/TimeModal.jsx
@@ -2,280 +2,11 @@
 import React, { useState } from 'react';
 // Libraries
 import PropTypes from 'prop-types';
-import Select from 'react-select';
-import moment from 'moment-timezone';
-// Components
+// Caseflow
 import Modal from '../../../components/Modal';
-// Styling
+import { TimeSelect } from '../scheduleHearing/TimeSelect';
+import { InfoAlert } from '../scheduleHearing/InfoAlert';
 import { css } from 'glamor';
-import { COLORS } from 'app/constants/AppConstants';
-
-/*
-* TIME LIST GENERATION START
-*/
-// Get an array with every fifteen minute increment in a day
-// [00:00, 00:15, ... , 23:45]
-const generateTimes = (roTimezone, date, intervalMinutes = 15) => {
-  // Start at midnight '00:00' is the first minute of a day
-  const currentTime = moment.tz(`${date} 00:00`, 'YYYY-MM-DD HH:mm', roTimezone);
-  // End at 23:59, the last minute of a day
-  const elevenFiftyNine = moment.tz(`${date} 23:59`, 'YYYY-MM-DD HH:mm', roTimezone);
-
-  const times = [];
-
-  // Go through the day in fifteen minute increments and store each increment
-  while (currentTime.isBefore(elevenFiftyNine)) {
-    // Moment has mutable objects so clone() is necessary
-    times.push(currentTime.clone());
-    currentTime.add(intervalMinutes, 'minute');
-  }
-
-  return times;
-};
-// Move the part of the arrawy after newFirstValue to the end of the array
-const moveTimesToEndOfArray = (newFirstValue, times) => {
-  // Find the index of newFirstValue
-  const firstValueIndex = times.findIndex((time) => time.isSame(newFirstValue));
-
-  // Remove all values before newFirstValue from the front of the array
-  const beforeFirstValue = times.slice(0, firstValueIndex);
-  const afterFirstValue = times.slice(firstValueIndex);
-
-  // Add the values back onto the end of the array
-  return afterFirstValue.concat(beforeFirstValue);
-};
-// Convert each time in the array into the expected 'option' format for react-select
-const formatTimesToOptionObjects = (times) => {
-  return times.map((time) => {
-    return {
-      label: time.format('h:mm A'),
-      value: time
-    };
-  });
-};
-// Generate a time for every 15m increment in a day.
-// Then move every time before beginsAt to the end of
-// the array to beginsAt appears first.
-const generateOrderedTimeOptions = (roTimezone, hearingDayDate) => {
-  const beginsAt = moment.tz(`${hearingDayDate} 08:30`, 'YYYY-MM-DD HH:mm', roTimezone);
-  const times = generateTimes(roTimezone, hearingDayDate);
-  const reorderedTimes = moveTimesToEndOfArray(beginsAt, times);
-  const options = formatTimesToOptionObjects(reorderedTimes);
-
-  return options;
-};
-
-/*
-* TIME LIST GENERATION END
-*/
-
-/*
-* CUSTOM LIST FILTERING START
-*/
-// Checks if the input matches the hour of a candidate.value which is a moment object
-const matchesHour = (candidate, input, exact = false) => {
-  const candidateHourString = candidate.value.format('h');
-
-  return exact ? candidateHourString === input : candidateHourString.startsWith(input);
-};
-const removeOneLeadingZero = (string) => {
-  return string[0] === '0' ? string.slice(1) : string;
-};
-// Checks if the input matches any part of a candidate.value which is a moment object
-const matchesAny = (candidate, input) => {
-  if (input.includes(':')) {
-    // Split into hours and minutes
-    const [hour, minutesAndAmPm] = input.split(':');
-
-    // Check that the hour matches exactly and the minutes+ampm are present
-    return matchesHour(candidate, hour, true) && matchesAny(candidate, minutesAndAmPm);
-  }
-  if (!input.includes(':')) {
-    // Produce a time like '400pm' or '800am' for string searching
-    const candidateNoColon = candidate.value.format('hhmmA');
-    // Remove spaces, force upper case so AM/PM searching works
-    const noColonOrSpaces = input.replace(' ', '').toUpperCase();
-    // Remove a leading zero if there is one
-    const noLeadingZero = removeOneLeadingZero(noColonOrSpaces);
-
-    return candidateNoColon.includes(noLeadingZero);
-  }
-};
-// Filter the options list to display only options that match
-// what's been typed into the input
-const filterOptions = (candidate, input) => {
-  // If only one character in the input assume it represents an hour
-  if (input.length === 1) {
-    return matchesHour(candidate, input);
-  }
-  // If one character and ':' in the input assume it represents an hour
-  if (input.length === 2 && input.endsWith(':')) {
-    return matchesHour(candidate, input[0], true);
-  }
-  // For everything else, send to matchesAny, which also handles ':'
-  if (input.length >= 2) {
-    return matchesAny(candidate, input);
-  }
-};
-
-/*
-* CUSTOM LIST FILTERING END
-*/
-// Given a long timezone like "America/Los_Angeles" return the
-// short version like "PDT" or "PST" (depending on date)
-const getTimezoneAbbreviation = (timezone) => {
-  // Create a moment object so we can extract the timezone
-  // abbreviation like 'PDT'
-  return moment.tz('00:00', 'HH:mm', timezone).format('z');
-};
-
-// Should maybe be made part of the "Alert" component? Seemed very different
-const InfoAlert = ({ timeString }) => {
-  const alertContainerStyles = css({
-    display: 'flex',
-    alignItems: 'center'
-  });
-
-  const greyRectangleStyles = css({
-    background: COLORS.GREY_LIGHT,
-    width: '1rem',
-    height: '4rem',
-    display: 'inline-block',
-    marginRight: '1.5rem'
-  });
-  const textDivStyles = css({
-    display: 'inline-block',
-    fontStyle: 'italic'
-  });
-
-  return (
-    <div className="info-alert" {...alertContainerStyles}>
-      <div {...greyRectangleStyles} />
-      <div {...textDivStyles}>{`The hearing will start at ${timeString} Eastern Time`}</div>
-    </div>
-  );
-};
-
-InfoAlert.propTypes = {
-  timeString: PropTypes.string
-};
-
-const TimeSelect = ({ roTimezone, onSelect, error, clearError, hearingDayDate }) => {
-
-  const customSelectStyles = {
-    // Hide the dropdown arrow on the right side
-    indicatorSeparator: () => ({ display: 'none' }),
-    dropdownIndicator: () => ({ display: 'none' }),
-    // Set the height of the select component
-    valueContainer: (styles) => ({
-      ...styles,
-      border: error ? '2px solid red' : styles.border,
-      height: '44px',
-      minHeight: '44px',
-    }),
-    // Fix selected text positioning problem caused by adjusting height
-    singleValue: () => ({
-      padding: '0',
-      margin: '0'
-    }),
-    // Change the highlight colors in the dropdown to gray
-    option: (styles, { isFocused }) => ({
-      ...styles,
-      color: isFocused ? 'white' : styles.color,
-      backgroundColor: isFocused ? COLORS.GREY : null,
-      ':hover': {
-        ...styles[':hover'],
-        backgroundColor: COLORS.GREY,
-        color: 'white'
-      },
-    })
-  };
-  const containerStyles = css({
-    borderRadius: '5px',
-    background: COLORS.GREY_LIGHT,
-    width: '50%',
-    marginTop: '16px',
-    marginBottom: '32px',
-    display: 'flex',
-    alignItems: 'center'
-  });
-  const selectContainerStyles = css({
-    width: '75%',
-    display: 'inline-block'
-  });
-  const timezoneAbbreviationContainer = css({
-    width: '25%',
-    display: 'inline-block',
-    textAlign: 'center',
-    color: 'black',
-    fontWeight: 'bold'
-  });
-
-  // This code exists to customize when the menu is shown/hidden
-  // our requirements have the menu ONLY shown when there's something
-  // entered in the input and nothing has been selected.
-  const [menuOpen, setMenuOpen] = useState(false);
-  const hideMenu = () => setMenuOpen(false);
-  const showMenu = () => setMenuOpen(true);
-  const handleInputChange = (query, { action }) => {
-    // Clear the error as soon as the input is interacted with
-    clearError();
-    // Show the menu if there are any characters in the select box
-    if (action === 'input-change' && query) {
-      showMenu();
-    }
-    // When deleting, if no characters in select box, hide the menu
-    if (action === 'input-change' && !query) {
-      hideMenu();
-    }
-  };
-
-  const options = generateOrderedTimeOptions(roTimezone, hearingDayDate);
-
-  return (
-    <div className="time-select" {...containerStyles}>
-      <div {...selectContainerStyles}>
-        <Select
-          // Accessibility
-          aria-label="time select"
-          // Make this a controlled select
-          onChange={onSelect}
-          // Backspace will clear a selected option, also show an 'x' to clear
-          isClearable
-          // Make this a searchable react-select
-          isSearchable
-          // The array of options
-          options={options}
-          // Custom searching logic
-          filterOption={filterOptions}
-          // Several options together to force the menu to be closed until typed into
-          onInputChange={handleInputChange}
-          menuIsOpen={menuOpen}
-          onBlur={hideMenu}
-          closeMenuOnSelect
-          blurInputOnSelect
-          // Dont show the placeholder text
-          placeholder=""
-          // Hide some elements of react-select, deal with error state, adjust height of component
-          styles={customSelectStyles}
-          // Set the maximum height
-          maxMenuHeight={175}
-        />
-      </div>
-      <div {...timezoneAbbreviationContainer}>
-        {getTimezoneAbbreviation(roTimezone)}
-      </div>
-    </div>
-  );
-};
-
-TimeSelect.propTypes = {
-  roTimezone: PropTypes.string,
-  onSelect: PropTypes.func,
-  error: PropTypes.string,
-  clearError: PropTypes.func,
-  hearingDayDate: PropTypes.string
-};
 
 export const TimeModal = ({ onCancel, onConfirm, ro, title, hearingDayDate }) => {
   // Error message state
